feat(package): [mock] extracted from @onr/core
