+++ packages/mock/package.json
@@ -1,8 +1,8 @@
 {
-  "name": "@onr/core",
-  "version": "0.3.2",
-  "description": "core module",
-  "author": "Raymond Tam <raymondtamtwtp@gmail.com>",
+  "name": "@onr/mock",
+  "version": "0.0.1",
+  "description": "mock module",
+  "author": "",
   "homepage": "https://github.com/OnrampLab/onr-react-ui#readme",
   "license": "ISC",
   "main": "lib/index.js",
@@ -30,26 +30,15 @@
     "url": "https://github.com/OnrampLab/onr-react-ui/issues"
   },
   "peerDependencies": {
-    "antd": "^4.16.1",
-    "lodash": "^4.14.0",
-    "next": "^10.0.3",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "react-feather": "^2.0.9",
-    "react-redux": "^7.2.2",
-    "redux": "^4.0.1",
-    "redux-thunk": "^2.3.0",
-    "styled-components": "^4.3.1"
+    "miragejs": "^0.1.41"
   },
   "devDependencies": {
-    "@babel/preset-typescript": "^7.12.7",
+    "@onr/common": "^0.4.0",
+    "@onr/eslint-config": "^0.2.3",
     "@onr/prettier-config": "^1.1.0",
-    "@types/antd": "^1.0.0",
-    "@types/lodash": "^4.14.168",
     "@types/node": "^14.14.21",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
-    "babel-plugin-module-resolver": "^4.1.0",
+    "eslint": "^7.19.0",
+    "prettier": "^1.18.2",
     "typescript": "^4.1.3"
   },
   "prettier": "@onr/prettier-config",
