+++ packages/mock/src/mock/Mock.ts
@@ -67,7 +67,7 @@ export class Mock implements IMock {
   private registerResourceRoutes(server: Server, resource: any) {
     debug(`register resource routes => ${resource}`);
 
-    server.get(`/${resource}`, function(schema) {
+    server.get(`/${resource}`, function (schema) {
       const data = this.serialize(schema[resource].all())[resource];
 
       return {
