+++ packages/mock/tsconfig.json
@@ -1,0 +1,10 @@
+{
+  "extends": "../../tsconfig.components.json",
+  "compilerOptions": {
+    "outDir": "lib",
+    "declarationDir": "./types",
+    "sourceMap": true
+  },
+  "include": ["src/**/*"],
+  "exclude": ["**/*.spec.tsx?"]
+}
