+++ packages/next-starter/package.json
@@ -26,6 +26,7 @@
     "@ant-design/icons": "^4.2.1",
     "@onr/common": "^0.4.0",
     "@onr/core": "^0.3.2",
+    "@onr/mock": "^0.0.1",
     "@tailwindcss/postcss7-compat": "^2.0.2",
     "@zeit/next-bundle-analyzer": "^0.1.2",
     "@zeit/next-less": "1.0.2-canary.2",
