+++ packages/next-starter/src/modules/auth/__mocks__/index.ts
@@ -1,4 +1,4 @@
-import { Mock } from '@onr/core';
+import { Mock } from '@onr/mock';
 import { IRole, roleModel } from './roleModel';
 import authRoutes from './routes.json';
 import roleSeeds from './roleSeeds.json';
