+++ packages/next-starter/src/pages/_app.tsx
@@ -1,5 +1,5 @@
 if (process.env.NODE_ENV === 'development' && process.env.NEXT_PUBLIC_ENABLE_MIRAGE === 'true') {
-  const { makeServer } = require('@onr/core');
+  const { makeServer } = require('@onr/mock');
   const { mocks } = require('@app/__mocks__');
 
   makeServer({ environment: 'development', mocks });
