+++ packages/onr-core/src/index.ts
@@ -1,5 +1,4 @@
 export * from './components';
 export * from './hooks';
-export * from './mock';
 export * from './redux';
 export * from './types';
